+++ org/address.go
@@ -32,7 +32,7 @@ type Address struct {
 	// Additional street address details.
 	StreetExtra string `json:"street_extra,omitempty" jsonschema:"title=Extended Street"`
 	// Village, town, district, or city, typically inside a region.
-	Locality string `json:"locality" jsonschema:"title=Locality"`
+	Locality string `json:"locality,omitempty" jsonschema:"title=Locality"`
 	// Province, county, or state, inside a country.
 	Region string `json:"region,omitempty" jsonschema:"title=Region"`
 	// Post or ZIP code.
@@ -54,7 +54,6 @@ func (a *Address) Validate() error {
 func (a *Address) ValidateWithContext(ctx context.Context) error {
 	return tax.ValidateStructWithRegime(ctx, a,
 		validation.Field(&a.UUID),
-		validation.Field(&a.Locality, validation.Required),
 		validation.Field(&a.Country),
 		validation.Field(&a.Coordinates),
 		validation.Field(&a.Meta),