+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
-## Upcoming
+## [Unreleased]
 
-- Describe upcoming changes here.
+- Coming soon...
 
 ## [v0.111.0] - 2024-07-24