+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [v0.111.1] - 2024-07-25
+
 ### Added
 
 - `org.Address`: recommended fields added